include/VBox/pci.h,Main,FE/VBoxManage: Some smaller cleanup for the PCIBusAddress class and let it take the PCI domain as well, enable and set the PCI express port type on the bridge automatically when configuring a passthrough device, bugref:11054
svn:sync-xref-src-repo-rev: r173196

+++ src/VBox/Main/src-all/PCIDeviceAttachmentImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: PCIDeviceAttachmentImpl.cpp 112403 2026-01-11 19:29:08Z knut.osmundsen@oracle.com $ */
+/* $Id: PCIDeviceAttachmentImpl.cpp 113442 2026-03-17 09:21:49Z alexander.eichner@oracle.com $ */
 /** @file
  * PCI attachment information implmentation.
  */
@@ -36,8 +36,8 @@
 struct PCIDeviceAttachment::Data
 {
     Data(const Utf8Str &aDevName,
-         LONG          aHostAddress,
-         LONG          aGuestAddress,
+         ULONG         aHostAddress,
+         ULONG         aGuestAddress,
          BOOL          afPhysical) :
         DevName(aDevName),
         HostAddress(aHostAddress),
@@ -47,8 +47,8 @@ struct PCIDeviceAttachment::Data
     }
 
     Utf8Str          DevName;
-    LONG             HostAddress;
-    LONG             GuestAddress;
+    ULONG            HostAddress;
+    ULONG            GuestAddress;
     BOOL             fPhysical;
 };
 
@@ -73,8 +73,8 @@ void PCIDeviceAttachment::FinalRelease()
 /////////////////////////////////////////////////////////////////////////////
 HRESULT PCIDeviceAttachment::init(IMachine      *aParent,
                                   const Utf8Str &aDevName,
-                                  LONG          aHostAddress,
-                                  LONG          aGuestAddress,
+                                  ULONG         aHostAddress,
+                                  ULONG         aGuestAddress,
                                   BOOL          fPhysical)
 {
     NOREF(aParent);
@@ -103,17 +103,15 @@ HRESULT PCIDeviceAttachment::initCopy(IMachine *aParent, PCIDeviceAttachment *aT
 HRESULT PCIDeviceAttachment::i_loadSettings(IMachine *aParent,
                                             const settings::HostPCIDeviceAttachment &hpda)
 {
-    /** @todo r=bird: Inconsistent signed/unsigned crap. */
-    return init(aParent, hpda.strDeviceName, (LONG)hpda.uHostAddress, (LONG)hpda.uGuestAddress, TRUE);
+    return init(aParent, hpda.strDeviceName, hpda.uHostAddress, hpda.uGuestAddress, TRUE);
 }
 
 
 HRESULT PCIDeviceAttachment::i_saveSettings(settings::HostPCIDeviceAttachment &data)
 {
     Assert(m);
-    /** @todo r=bird: Inconsistent signed/unsigned crap. */
-    data.uHostAddress  = (uint32_t)m->HostAddress;
-    data.uGuestAddress = (uint32_t)m->GuestAddress;
+    data.uHostAddress  = m->HostAddress;
+    data.uGuestAddress = m->GuestAddress;
     data.strDeviceName = m->DevName;
 
     return S_OK;
@@ -148,12 +146,12 @@ HRESULT PCIDeviceAttachment::getIsPhysicalDevice(BOOL *aIsPhysicalDevice)
     return S_OK;
 }
 
-HRESULT PCIDeviceAttachment::getHostAddress(LONG *aHostAddress)
+HRESULT PCIDeviceAttachment::getHostAddress(ULONG *aHostAddress)
 {
     *aHostAddress = m->HostAddress;
     return S_OK;
 }
-HRESULT PCIDeviceAttachment::getGuestAddress(LONG *aGuestAddress)
+HRESULT PCIDeviceAttachment::getGuestAddress(ULONG *aGuestAddress)
 {
     *aGuestAddress = m->GuestAddress;
     return S_OK;
